Changed updated_by to a bigint

+++ db/migrate/20210830135849_add_updated_by_to_sports.rb
@@ -1,5 +1,5 @@
 class AddUpdatedByToSports < ActiveRecord::Migration[6.1]
   def change
-    add_column :sports, :updated_by, :integer
+    add_column :sports, :updated_by, :bigint
   end
 end
